+++ PetLab.WPF/Models/OrderViewModel.cs
@@ -9,7 +9,7 @@ namespace PetLab.WPF.Models {
 		public int ShiftId { get; set; }
 		public string DyeName { get; set; }
 		public string ColorShade { get; set; }
-		public short CountSocket { get; set; }
+		public byte CountSocket { get; set; }
 		public string EquipmentId { get; set; }
 
 		public OrderEtalonColorViewModel EtalonColor {
